requirements.txt
streamlit
matplotlib
numpy
scipy
pillow
opencv-python-headless
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fonttools==4.59.2
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.6
narwhals==2.4.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.3.0
protobuf==6.32.0
pyarrow==21.0.0
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.8.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
streamlit==1.49.1
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.5.10
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
tensorflow>=2.16.0
tabulate>=0.8.9
tqdm==4.66.4
plotly==5.24.1
PyMuPDF==1.24.9
